requirements.txt
streamlit
openai
pandas
boto3
python-dotenv
